feat(header): update navigation header
