+++ client/src/layout/components/index.ts
@@ -5,7 +5,6 @@ import Header from 'layout/components/Header'
 import DomainHeader from 'layout/components/DomainHeader'
 import NotFound from 'layout/components/NotFound'
 import HeaderBackground from 'layout/components/HeaderBackground'
-import HeaderDropdownMenu from 'layout/components/HeaderDropdownMenu'
 import HeaderChainDropdown from 'layout/components/HeaderChainDropdown'
 import MobileHeader from 'layout/components/MobileHeader'
 import ExtrinsicAndEventResultTable from 'layout/components/ExtrinsicAndEventResult'
@@ -18,7 +17,6 @@ export {
   DomainHeader,
   NotFound,
   HeaderBackground,
-  HeaderDropdownMenu,
   HeaderChainDropdown,
   MobileHeader,
   ExtrinsicAndEventResultTable,
